search screen no SQL
Fixed a few errors with screen switching, probably a better way to do that but it works for now. Added search screen and general functionality, but not any SQL commands. The special character function tests against ALL special chars not just *, so we may need to do something about that.